feat(sdk): move `bump_event_types` to the list sub-request
