+++ bindings/matrix-sdk-ffi/src/sliding_sync.rs
@@ -560,6 +560,14 @@ impl SlidingSyncListBuilder {
         );
         Arc::new(builder)
     }
+
+    pub fn bump_event_types(self: Arc<Self>, bump_event_types: Vec<String>) -> Arc<Self> {
+        let mut builder = unwrap_or_clone_arc(self);
+        builder.inner = builder.inner.bump_event_types(
+            bump_event_types.into_iter().map(Into::into).collect::<Vec<_>>().as_slice(),
+        );
+        Arc::new(builder)
+    }
 }
 
 pub trait SlidingSyncListOnceBuilt: Sync + Send {
@@ -873,14 +881,6 @@ impl SlidingSyncBuilder {
     pub fn with_all_extensions(self: Arc<Self>) -> Arc<Self> {
         let mut builder = unwrap_or_clone_arc(self);
         builder.inner = builder.inner.with_all_extensions();
-        Arc::new(builder)
-    }
-
-    pub fn bump_event_types(self: Arc<Self>, bump_event_types: Vec<String>) -> Arc<Self> {
-        let mut builder = unwrap_or_clone_arc(self);
-        builder.inner = builder.inner.bump_event_types(
-            bump_event_types.into_iter().map(Into::into).collect::<Vec<_>>().as_slice(),
-        );
         Arc::new(builder)
     }
