+++ crates/matrix-sdk/src/sliding_sync/list/builder.rs
@@ -10,7 +10,11 @@ use std::{
 use eyeball::unique::Observable;
 use eyeball_im::ObservableVector;
 use imbl::Vector;
-use ruma::{api::client::sync::sync_events::v4, events::StateEventType, OwnedRoomId};
+use ruma::{
+    api::client::sync::sync_events::v4,
+    events::{StateEventType, TimelineEventType},
+    OwnedRoomId,
+};
 use tokio::sync::broadcast::Sender;
 
 use super::{
@@ -53,6 +57,8 @@ pub struct SlidingSyncListBuilder {
     reloaded_cached_data: Option<SlidingSyncListCachedData>,
 
     once_built: Arc<Box<dyn Fn(SlidingSyncList) -> SlidingSyncList + Send + Sync>>,
+
+    bump_event_types: Vec<TimelineEventType>,
 }
 
 // Print debug values for the builder, except `once_built` which is ignored.
@@ -66,6 +72,7 @@ impl fmt::Debug for SlidingSyncListBuilder {
             .field("filters", &self.filters)
             .field("timeline_limit", &self.timeline_limit)
             .field("name", &self.name)
+            .field("bump_event_types", &self.bump_event_types)
             .finish_non_exhaustive()
     }
 }
@@ -85,6 +92,7 @@ impl SlidingSyncListBuilder {
             reloaded_cached_data: None,
             cache_policy: SlidingSyncListCachePolicy::Disabled,
             once_built: Arc::new(Box::new(identity)),
+            bump_event_types: Vec::new(),
         }
     }
 
@@ -166,6 +174,19 @@ impl SlidingSyncListBuilder {
         }
     }
 
+    /// Allowlist of event types which should be considered recent activity
+    /// when sorting `by_recency`.
+    ///
+    /// By omitting event types, clients can ensure
+    /// that uninteresting events (e.g. a profile rename) do not cause a
+    /// room to jump to the top of its list(s). Empty or
+    /// omitted `bump_event_types` have no effect: all events in a room will
+    /// be considered recent activity.
+    pub fn bump_event_types(mut self, bump_event_types: &[TimelineEventType]) -> Self {
+        self.bump_event_types = bump_event_types.to_vec();
+        self
+    }
+
     /// Build the list.
     pub(in super::super) fn build(
         self,
@@ -180,6 +201,7 @@ impl SlidingSyncListBuilder {
                 timeline_limit: StdRwLock::new(self.timeline_limit),
                 name: self.name,
                 cache_policy: self.cache_policy,
+                bump_event_types: self.bump_event_types,
 
                 // Computed from the builder.
                 request_generator: StdRwLock::new(SlidingSyncListRequestGenerator::new(
