+++ crates/matrix-sdk/src/sliding_sync/list/mod.rs
@@ -18,7 +18,12 @@ pub(super) use frozen::FrozenSlidingSyncList;
 use futures_core::Stream;
 pub(super) use request_generator::*;
 pub use room_list_entry::RoomListEntry;
-use ruma::{api::client::sync::sync_events::v4, assign, events::StateEventType, OwnedRoomId};
+use ruma::{
+    api::client::sync::sync_events::v4,
+    assign,
+    events::{StateEventType, TimelineEventType},
+    OwnedRoomId,
+};
 use serde::{Deserialize, Serialize};
 use tokio::sync::broadcast::Sender;
 use tracing::{instrument, warn};
@@ -251,6 +256,10 @@ pub(super) struct SlidingSyncListInner {
     /// The Sliding Sync internal channel sender. See
     /// [`SlidingSyncInner::internal_channel`] to learn more.
     sliding_sync_internal_channel_sender: Sender<SlidingSyncInternalMessage>,
+
+    /// The `bump_event_types` field. See
+    /// [`SlidingSyncListBuilder::bump_event_types`] to learn more.
+    bump_event_types: Vec<TimelineEventType>,
 }
 
 impl SlidingSyncListInner {
@@ -314,6 +323,7 @@ impl SlidingSyncListInner {
             }),
             sort,
             filters,
+            bump_event_types: self.bump_event_types.clone(),
         })
     }
